コードの分割。
